requirements-cdk.txt
aws-cdk-lib==2.117.0
constructs>=10.0.0,<11.0.0
boto3==1.26.137
